Move arrow into sub-package.
